requirements.txt
Flask
requests
Flask-Limiter
python-dotenv
redis